Struct leo_passes::common::assigner::Assigner

pub struct Assigner {
    pub(crate) counter: usize,
}

A struct used to create assignment statements.

Fields
counter: usize

A strictly increasing counter, used to ensure that new variable names are unique.

Implementations

impl Assigner

pub(crate) fn unique_symbol(
&mut self,
arg: impl Display,
separator: impl Display
) -> Symbol

Return a new unique Symbol from a &str.

pub(crate) fn simple_assign_statement(
&mut self,
identifier: Identifier,
value: Expression
) -> Statement

Constructs the assignment statement place = expr;. This function should be the only place where AssignStatements are constructed.

pub(crate) fn unique_simple_assign_statement(
&mut self,
expr: Expression
) -> (Identifier, Statement)

Constructs a simple assign statement for expr with a unique name. For example, expr is transformed into $var$0 = expr;.