leo_passes::common::graph

Type Alias StructGraph

pub type StructGraph = DiGraph<Symbol>;

A struct dependency graph.

Aliased Type

struct StructGraph {
    nodes: IndexSet<Symbol>,
    edges: IndexMap<Symbol, IndexSet<Symbol>>,
}

Fields
nodes: IndexSet<Symbol>

The set of nodes in the graph.

edges: IndexMap<Symbol, IndexSet<Symbol>>

The directed edges in the graph. Each entry in the map is a node in the graph, and the set of nodes that it points to.