leo_passes::common
Module symbol_table

Modules

symbols 🔒

Structs

Finalizer

FunctionSymbol

LocalTable 🔒

LocalTableInner 🔒

SymbolTable

Maps global and local symbols to information about them.

VariableSymbol

An entry for a variable in the symbol table.

Enums

VariableType

An enumeration of the different types of variable type.

Functions

eq_struct 🔒