leo_passes::common_subexpression_elimination::visitor

Enum Expr

pub enum Expr {
    Atom(Atom),
    Array(Vec<Atom>),
    ArrayAccess {
        array: Atom,
        index: Atom,
    },
    Binary {
        op: BinaryOperation,
        left: Atom,
        right: Atom,
    },
    Repeat {
        value: Atom,
        count: Atom,
    },
    Ternary {
        condition: Atom,
        if_true: Atom,
        if_false: Atom,
    },
    Unary {
        op: UnaryOperation,
        receiver: Atom,
    },
}

An expression that can be mapped to a variable, and eliminated if it appears again.

For now we are rather conservative in the types of expressions we allow. We define this separate type rather than using Expression largely for ease of hashing and comparison while ignoring superfluous information like node ids and spans. It also makes explicit in the type that subexpressions must be atoms.

Variants

Atom(Atom)

Array(Vec<Atom>)

ArrayAccess

Fields

array: Atom

index: Atom

Binary

Fields

op: BinaryOperation

left: Atom

right: Atom

Repeat

Fields

value: Atom

count: Atom

Ternary

Fields

condition: Atom

if_true: Atom

if_false: Atom

Unary

Fields

op: UnaryOperation

receiver: Atom