Struct leo_passes::flattening::flattener::Flattener

pub struct Flattener<'a> {
    pub(crate) symbol_table: &'a SymbolTable,
    pub(crate) type_table: &'a TypeTable,
    pub(crate) node_builder: &'a NodeBuilder,
    pub(crate) assigner: &'a Assigner,
    pub(crate) condition_stack: Vec<Expression>,
    pub(crate) returns: Vec<(Option<Expression>, ReturnStatement)>,
    pub(crate) program: Option<Symbol>,
}

Fields
symbol_table: &'a SymbolTable

The symbol table associated with the program.

type_table: &'a TypeTable

A mapping between node IDs and their types.

node_builder: &'a NodeBuilder

A counter used to generate unique node IDs.

assigner: &'a Assigner

A struct used to construct (unique) assignment statements.

condition_stack: Vec<Expression>

A stack of condition Expressions visited up to the current point in the AST.

returns: Vec<(Option<Expression>, ReturnStatement)>

A list containing tuples of guards and expressions associated ReturnStatements. A guard is an expression that evaluates to true on the execution path of the ReturnStatement. Note that returns are inserted in the order they are encountered during a pre-order traversal of the AST. Note that type checking guarantees that there is at most one return in a basic block.

program: Option<Symbol>

The program name.

Implementations

impl<'a> Flattener<'a>

pub(crate) fn new( symbol_table: &'a SymbolTable, type_table: &'a TypeTable, node_builder: &'a NodeBuilder, assigner: &'a Assigner ) -> Self

pub(crate) fn clear_early_returns( &mut self ) -> Vec<(Option<Expression>, ReturnStatement)>

Clears the state associated with ReturnStatements, returning the ones that were previously stored.

pub(crate) fn construct_guard(&mut self) -> Option<Expression>

Constructs a guard from the current state of the condition stack.

pub(crate) fn fold_guards( &mut self, prefix: &str, guards: Vec<(Option<Expression>, Expression)> ) -> (Expression, Vec<Statement>)

Fold guards and expressions into a single expression. Note that this function assumes that at least one guard is present.

pub(crate) fn unique_simple_assign_statement( &mut self, expr: Expression ) -> (Identifier, Statement)

A wrapper around assigner.unique_simple_assign_statement that updates self.structs.

pub(crate) fn simple_assign_statement( &mut self, lhs: Identifier, rhs: Expression ) -> Statement

A wrapper around assigner.simple_assign_statement that tracks the type of the lhs.

pub(crate) fn fold_returns( &mut self, block: &mut Block, returns: Vec<(Option<Expression>, ReturnStatement)> )

Folds a list of return statements into a single return statement and adds the produced statements to the block.

pub(crate) fn ternary_array( &mut self, array: &ArrayType, condition: &Expression, first: &Identifier, second: &Identifier ) -> (Expression, Vec<Statement>)

pub(crate) fn ternary_struct( &mut self, struct_: &Composite, condition: &Expression, first: &Identifier, second: &Identifier ) -> (Expression, Vec<Statement>)

pub(crate) fn ternary_tuple( &mut self, tuple_type: &TupleType, condition: &Expression, first: &Identifier, second: &Identifier ) -> (Expression, Vec<Statement>)