leo_passes::flattening::visitor

Struct FlatteningVisitor

pub struct FlatteningVisitor<'a> {
    pub state: &'a mut CompilerState,
    pub condition_stack: Vec<Guard>,
    pub returns: Vec<(ReturnGuard, ReturnStatement)>,
    pub program: Symbol,
    pub is_async: bool,
}

Fields
state: &'a mut CompilerStatecondition_stack: Vec<Guard>

A stack of condition Expressions visited up to the current point in the AST.

returns: Vec<(ReturnGuard, ReturnStatement)>

A list containing tuples of guards and expressions associated ReturnStatements. A guard is an expression that evaluates to true on the execution path of the ReturnStatement. Note that returns are inserted in the order they are encountered during a pre-order traversal of the AST. Note that type checking guarantees that there is at most one return in a basic block.

program: Symbol

The program name.

is_async: bool

Whether the function is an async function.

Implementations

impl FlatteningVisitor<'_>

pub fn construct_early_return_guard( &mut self, ) -> Option<(Identifier, Vec<Statement>)>

Construct an early return guard.

That is, an Identifier assigned to a boolean that is true iff some early return was taken.

pub fn construct_guard(&mut self) -> Option<(Identifier, Vec<Statement>)>

Construct a guard from the current state of the condition stack.

That is, a boolean expression which is true iff we’ve followed the branches that led to the current point in the Leo code.

pub fn fold_guards( &mut self, prefix: &str, guards: Vec<(Option<Expression>, Expression)>, ) -> (Expression, Vec<Statement>)

Fold guards and expressions into a single expression. Note that this function assumes that at least one guard is present.

pub fn unique_simple_definition( &mut self, expr: Expression, ) -> (Identifier, Statement)

A wrapper around assigner.unique_simple_definition that updates self.structs.

pub fn simple_definition( &mut self, lhs: Identifier, rhs: Expression, ) -> Statement

A wrapper around assigner.simple_definition that tracks the type of the lhs.

pub fn fold_returns( &mut self, block: &mut Block, returns: Vec<(Option<Expression>, ReturnStatement)>, )

Folds a list of return statements into a single return statement and adds the produced statements to the block.

fn make_array_access_definition( &mut self, i: usize, identifier: Identifier, array_type: &ArrayType, ) -> (Identifier, Statement)

pub fn ternary_array( &mut self, array: &ArrayType, condition: &Expression, first: &Identifier, second: &Identifier, ) -> (Expression, Vec<Statement>)

fn make_struct_access_definition( &mut self, inner: Identifier, name: Identifier, type_: Type, ) -> (Identifier, Statement)

pub fn ternary_struct( &mut self, struct_path: &Path, struct_: &Composite, condition: &Expression, first: &Identifier, second: &Identifier, ) -> (Expression, Vec<Statement>)

pub fn ternary_tuple( &mut self, tuple_type: &TupleType, condition: &Expression, first: &Expression, second: &Expression, ) -> (Expression, Vec<Statement>)