Struct leo_passes::function_inlining::function_inliner::FunctionInliner

pub struct FunctionInliner<'a> {
    pub(crate) call_graph: &'a CallGraph,
    pub(crate) assignment_renamer: AssignmentRenamer,
    pub(crate) reconstructed_functions: IndexMap<Symbol, Function>,
}

Fields
call_graph: &'a CallGraph

The call graph for the program.

assignment_renamer: AssignmentRenamer

A wrapper around an Assigner used to create unique variable assignments.

reconstructed_functions: IndexMap<Symbol, Function>

A map of reconstructed functions in the current program scope.

Implementations

impl<'a> FunctionInliner<'a>

pub fn new(call_graph: &'a CallGraph, assigner: Assigner) -> Self

Initializes a new FunctionInliner.