leo_passes::loop_unrolling::unroller
Struct Unroller

pub struct Unroller<'a> {
    pub(crate) symbol_table: &'a mut SymbolTable,
    pub(crate) type_table: &'a TypeTable,
    pub(crate) handler: &'a Handler,
    pub(crate) node_builder: &'a NodeBuilder,
    pub(crate) is_unrolling: bool,
    pub(crate) current_program: Option<Symbol>,
    pub(crate) loop_not_unrolled: Option<Span>,
    pub(crate) loop_unrolled: bool,
}

Fields
symbol_table: &'a mut SymbolTable

The symbol table for the function being processed.

type_table: &'a TypeTable

A mapping from node IDs to their types.

handler: &'a Handler

An error handler used for any errors found during unrolling.

node_builder: &'a NodeBuilder

A counter used to generate unique node IDs.

is_unrolling: bool

Are we in the midst of unrolling a loop?

current_program: Option<Symbol>

The current program name.

loop_not_unrolled: Option<Span>

If we’ve encountered a loop that was not unrolled, here’s it’s spanned.

loop_unrolled: bool

Have we unrolled any loop?

Implementations

impl<'a> Unroller<'a>

pub(crate) fn new( symbol_table: &'a mut SymbolTable, type_table: &'a TypeTable, handler: &'a Handler, node_builder: &'a NodeBuilder, ) -> Self

pub(crate) fn in_scope<T>( &mut self, id: NodeID, func: impl FnOnce(&mut Self) -> T, ) -> T

pub(crate) fn emit_err(&self, err: LoopUnrollerError)

Emits a Loop Unrolling Error

pub(crate) fn unroll_iteration_statement<I: LoopBound>( &mut self, input: IterationStatement, ) -> Statement

Unrolls an IterationStatement.

fn unroll_single_iteration<I: LoopBound>( &mut self, input: &IterationStatement, iteration_count: I, ) -> Statement

A helper function to unroll a single iteration an IterationStatement.