leo_passes::loop_unrolling::visitor

Struct UnrollingVisitor

pub struct UnrollingVisitor<'a> {
    pub state: &'a mut CompilerState,
    pub program: Symbol,
    pub loop_not_unrolled: Option<Span>,
    pub loop_unrolled: bool,
}

Fields
state: &'a mut CompilerStateprogram: Symbol

The current program name.

loop_not_unrolled: Option<Span>

If we’ve encountered a loop that was not unrolled, here’s it’s spanned.

loop_unrolled: bool

Have we unrolled any loop?

Implementations

impl UnrollingVisitor<'_>

pub fn in_scope<T>( &mut self, id: NodeID, func: impl FnOnce(&mut Self) -> T, ) -> T

pub fn emit_err(&self, err: LoopUnrollerError)

Emits a Loop Unrolling Error

pub fn unroll_iteration_statement<I: LoopBound>( &mut self, input: IterationStatement, start: Value, stop: Value, ) -> Statement

Unrolls an IterationStatement.

fn unroll_single_iteration<I: LoopBound>( &mut self, input: &IterationStatement, iteration_count: I, ) -> Statement

A helper function to unroll a single iteration an IterationStatement.