leo_passes::static_analysis::visitor

Struct StaticAnalyzingVisitor

pub struct StaticAnalyzingVisitor<'a> {
    pub state: &'a mut CompilerState,
    pub await_checker: AwaitChecker,
    pub current_program: Symbol,
    pub variant: Option<Variant>,
    pub non_async_external_call_seen: bool,
}

Fields
state: &'a mut CompilerStateawait_checker: AwaitChecker

Struct to store the state relevant to checking all futures are awaited.

current_program: Symbol

The current program name.

variant: Option<Variant>

The variant of the function that we are currently traversing.

non_async_external_call_seen: bool

Whether or not a non-async external call has been seen in this function.

Implementations

impl StaticAnalyzingVisitor<'_>

pub fn emit_err(&self, err: StaticAnalyzerError)

pub fn emit_warning(&self, warning: StaticAnalyzerWarning)

Emits a type checker warning

pub fn assert_future_await(&mut self, future: &Option<&Expression>, span: Span)

Type checks the awaiting of a future.

pub fn assert_simple_async_transition_call( &mut self, program: Symbol, function_path: &Path, span: Span, )

Assert that an async call is a “simple” one. Simple is defined as an async transition function which does not return a Future that itself takes a Future as an argument.