Struct leo_passes::static_single_assignment::static_single_assigner::StaticSingleAssigner

pub struct StaticSingleAssigner<'a> {
    pub(crate) symbol_table: &'a SymbolTable,
    pub(crate) rename_table: RenameTable,
    pub(crate) is_lhs: bool,
    pub(crate) assigner: Assigner,
}

Fields
symbol_table: &'a SymbolTable

The SymbolTable of the program.

rename_table: RenameTable

The RenameTable for the current basic block in the AST

is_lhs: bool

A flag to determine whether or not the traversal is on the left-hand side of a definition or an assignment.

assigner: Assigner

An struct used to construct (unique) assignment statements.

Implementations

impl<'a> StaticSingleAssigner<'a>

pub(crate) fn new(symbol_table: &'a SymbolTable) -> Self

Initializes a new StaticSingleAssigner with an empty RenameTable.

pub(crate) fn push(&mut self)

Pushes a new scope, setting the current scope as the new scope’s parent.

pub(crate) fn pop(&mut self) -> RenameTable

If the RenameTable has a parent, then self.rename_table is set to the parent, otherwise it is set to a default RenameTable.