leo_passes::static_single_assignment::visitor

Struct SsaFormingVisitor

pub struct SsaFormingVisitor<'a> {
    pub state: &'a mut CompilerState,
    pub rename_table: RenameTable,
    pub program: Symbol,
    pub rename_defs: bool,
}

Fields
state: &'a mut CompilerStaterename_table: RenameTable

The RenameTable for the current basic block in the AST

program: Symbol

The main program name.

rename_defs: bool

Whether to rename places in definitions.

Implementations

impl SsaFormingVisitor<'_>

pub(crate) fn push(&mut self)

Pushes a new scope, setting the current scope as the new scope’s parent.

pub(crate) fn pop(&mut self) -> RenameTable

If the RenameTable has a parent, then self.rename_table is set to the parent, otherwise it is set to a default RenameTable.

pub(crate) fn rename_identifier(&mut self, identifier: Identifier) -> Identifier

pub(crate) fn simple_definition( &mut self, identifier: Identifier, rhs: Expression, ) -> Statement

pub(crate) fn unique_simple_definition( &mut self, expr: Expression, ) -> (Identifier, Statement)

Constructs a simple assign statement for expr with a unique name. For example, expr is transformed into $var$0 = expr;. The lhs is guaranteed to be unique with respect to the Assigner.