Struct leo_passes::symbol_table_creator::creator::SymbolTableCreator

pub struct SymbolTableCreator<'a> {
    pub(crate) symbol_table: SymbolTable,
    handler: &'a Handler,
}

A compiler pass during which the SymbolTable is created. Note that this pass only creates the initial entries for functions, structs, and records. The table is populated further during the type checking pass.

Fields
symbol_table: SymbolTable

The SymbolTable constructed by this compiler pass.

handler: &'a Handler

The error handler.

Implementations

impl<'a> SymbolTableCreator<'a>

pub fn new(handler: &'a Handler) -> Self