leo_passes::type_checking::visitor

Struct TypeCheckingVisitor

pub struct TypeCheckingVisitor<'a> {
    pub state: &'a mut CompilerState,
    pub scope_state: ScopeState,
    pub async_function_input_types: IndexMap<Location, Vec<Type>>,
    pub async_function_callers: IndexMap<Location, IndexSet<Location>>,
    pub used_structs: IndexSet<Symbol>,
    pub limits: TypeCheckingInput,
    pub conditional_scopes: Vec<IndexSet<Symbol>>,
    pub async_block_id: Option<NodeID>,
}

Fields
state: &'a mut CompilerStatescope_state: ScopeState

The state of the current scope being traversed.

async_function_input_types: IndexMap<Location, Vec<Type>>

Mapping from async function stub name to the inferred input types.

async_function_callers: IndexMap<Location, IndexSet<Location>>

Mapping from async function name to the names of async transition callers.

used_structs: IndexSet<Symbol>

The set of used composites.

limits: TypeCheckingInput

So we can check if we exceed limits on array size, number of mappings, or number of functions.

conditional_scopes: Vec<IndexSet<Symbol>>

For detecting the error TypeCheckerError::async_cannot_assign_outside_conditional.

async_block_id: Option<NodeID>

If we’re inside an async block, this is the node ID of the contained Block. Otherwise, this is None.

Implementations

impl TypeCheckingVisitor<'_>

pub fn visit_expression_assign(&mut self, input: &Expression) -> Type

pub fn visit_array_access_general( &mut self, input: &ArrayAccess, assign: bool, expected: &Option<Type>, ) -> Type

pub fn visit_member_access_general( &mut self, input: &MemberAccess, assign: bool, expected: &Option<Type>, ) -> Type

pub fn visit_tuple_access_general( &mut self, input: &TupleAccess, assign: bool, expected: &Option<Type>, ) -> Type

pub fn visit_identifier_assign(&mut self, input: &Identifier) -> Type

pub(crate) fn visit_expression_reject_numeric( &mut self, expr: &Expression, expected: &Option<Type>, ) -> Type

Infers the type of an expression, but returns Type::Err and emits an error if the result is Type::Numeric. Used to disallow numeric types in specific contexts where they are not valid or expected.

pub(crate) fn visit_expression_infer_default_u32( &mut self, expr: &Expression, ) -> Type

Infers the type of an expression, and if it is Type::Numeric, coerces it to U32, validates it, and records it in the type table.

impl TypeCheckingVisitor<'_>

pub fn in_scope<T>( &mut self, id: NodeID, func: impl FnOnce(&mut Self) -> T, ) -> T

pub fn in_conditional_scope<T>( &mut self, func: impl FnOnce(&mut Self) -> T, ) -> T

pub fn insert_symbol_conditional_scope(&mut self, symbol: Symbol)

pub fn symbol_in_conditional_scope(&mut self, symbol: Symbol) -> bool

pub fn emit_err(&self, err: TypeCheckerError)

Emits a type checker error.

pub fn emit_warning(&mut self, warning: TypeCheckerWarning)

Emits a type checker warning

pub fn check_eq_types(&self, t1: &Option<Type>, t2: &Option<Type>, span: Span)

Emits an error if the two given types are not equal.

pub fn assert_and_return_type( &mut self, actual: Type, expected: &Option<Type>, span: Span, ) -> Type

Use this method when you know the actual type. Emits an error to the handler if the actual type is not equal to the expected type.

pub fn maybe_assert_type( &mut self, actual: &Type, expected: &Option<Type>, span: Span, )

pub fn assert_type(&mut self, actual: &Type, expected: &Type, span: Span)

pub fn assert_int_type(&self, type_: &Type, span: Span)

pub fn assert_unsigned_type(&self, type_: &Type, span: Span)

pub fn assert_bool_int_type(&self, type_: &Type, span: Span)

pub fn assert_field_int_type(&self, type_: &Type, span: Span)

pub fn assert_field_group_int_type(&self, type_: &Type, span: Span)

pub fn get_core_constant( &self, type_: &Type, constant: &Identifier, ) -> Option<CoreConstant>

Type checks the inputs to an associated constant and returns the expected output type.

pub fn get_core_function_call( &self, struct_: &Identifier, function: &Identifier, ) -> Option<CoreFunction>

Emits an error if the struct is not a core library struct. Emits an error if the function is not supported by the struct.

pub fn check_core_function_call( &mut self, core_function: CoreFunction, arguments: &[(Type, &Expression)], function_span: Span, ) -> Type

Type checks the inputs to a core function call and returns the expected output type. Emits an error if the correct number of arguments are not provided. Emits an error if the arguments are not of the correct type.

pub fn assert_member_is_not_record( &mut self, span: Span, parent: Symbol, type_: &Type, )

Emits an error if the struct member is a record type.

pub fn assert_type_is_valid(&mut self, type_: &Type, span: Span)

Emits an error if the type or its constituent types is not valid.

pub fn assert_mapping_type(&self, type_: &Type, span: Span)

Emits an error if the type is not a mapping.

pub fn assert_array_type(&self, type_: &Type, span: Span)

pub fn check_function_signature(&mut self, function: &Function, is_stub: bool)

Helper function to check that the input and output of function are valid

fn merge_types(&self, lhs: &mut Type, rhs: &Type)

Merge inferred types into lhs.

That is, if lhs and rhs aren’t equal, set lhs to Type::Err; or, if they’re both futures, set any member of lhs that isn’t equal to the equivalent member of rhs to Type::Err.

pub fn eq_user(&self, t1: &Type, t2: &Type) -> bool

Are the types considered equal as far as the Leo user is concerned?

In particular, any comparison involving an Err is true, composite types are resolved to the current program if not specified, and Futures which aren’t explicit compare equal to other Futures.

An array with an undetermined length (e.g., one that depends on a const) is considered equal to other arrays if their element types match. This allows const propagation to potentially resolve the length before type checking is performed again.

pub fn lookup_struct( &mut self, program: Option<Symbol>, name: Symbol, ) -> Option<Composite>

Wrapper around lookup_struct that additionally records all structs that are used in the program.

pub fn insert_variable( &mut self, inferred_type: Option<Type>, name: &Identifier, type_: Type, span: Span, )

Inserts variable to symbol table.

pub fn check_access_allowed( &mut self, name: &str, finalize_op: bool, span: Span, )

pub fn is_external_record(&self, ty: &Type) -> bool

pub fn parse_integer_literal<I: FromStrRadix>( &self, raw_string: &str, span: Span, type_string: &str, )

pub fn emit_inference_failure_error(&self, ty: &mut Type, expr: &Expression)

pub fn check_numeric_literal(&self, input: &Literal, ty: &Type) -> bool