Type Alias leo_passes::common::graph::StructGraph

pub type StructGraph = DiGraph<Symbol>;

A struct dependency graph.

Aliased Type

struct StructGraph {
    nodes: IndexSet<Symbol, RandomState>,
    edges: IndexMap<Symbol, IndexSet<Symbol, RandomState>, RandomState>,
}

Fields
nodes: IndexSet<Symbol, RandomState>

The set of nodes in the graph.

edges: IndexMap<Symbol, IndexSet<Symbol, RandomState>, RandomState>

The directed edges in the graph. Each entry in the map is a node in the graph, and the set of nodes that it points to.

Implementations

impl<N: Node> DiGraph<N>

pub fn new(nodes: IndexSet<N>) -> Self

Initializes a new DiGraph from a vector of source nodes.

pub fn add_edge(&mut self, from: N, to: N)

Adds an edge to the graph.

pub fn contains_node(&self, node: N) -> bool

Returns true if the graph contains the given node.

pub fn post_order(&self) -> Result<IndexSet<N>, DiGraphError<N>>

Returns the post-order ordering of the graph. Detects if there is a cycle in the graph.

fn contains_cycle_from( &self, node: N, discovered: &mut IndexSet<N>, finished: &mut IndexSet<N> ) -> Option<N>