Struct leo_passes::flattening::flattener::Flattener

pub struct Flattener<'a> {
    pub(crate) symbol_table: &'a SymbolTable,
    pub(crate) assigner: Assigner,
    pub(crate) structs: IndexMap<Symbol, Symbol>,
    pub(crate) condition_stack: Vec<Expression>,
    pub(crate) returns: Vec<(Option<Expression>, ReturnStatement)>,
    pub(crate) tuples: IndexMap<Symbol, TupleExpression>,
}

Fields
symbol_table: &'a SymbolTable

The symbol table associated with the program.

assigner: Assigner

An struct used to construct (unique) assignment statements.

structs: IndexMap<Symbol, Symbol>

The set of variables that are structs.

condition_stack: Vec<Expression>

A stack of condition Expressions visited up to the current point in the AST.

returns: Vec<(Option<Expression>, ReturnStatement)>

A list containing tuples of guards and expressions associated ReturnStatements. A guard is an expression that evaluates to true on the execution path of the ReturnStatement. Note that returns are inserted in the order they are encountered during a pre-order traversal of the AST. Note that type checking guarantees that there is at most one return in a basic block.

tuples: IndexMap<Symbol, TupleExpression>

A mapping between variables and flattened tuple expressions.

Implementations

impl<'a> Flattener<'a>

pub(crate) fn new(symbol_table: &'a SymbolTable, assigner: Assigner) -> Self

pub(crate) fn clear_early_returns(
&mut self
) -> Vec<(Option<Expression>, ReturnStatement)>

Clears the state associated with ReturnStatements, returning the ones that were previously stored.

pub(crate) fn construct_guard(&mut self) -> Option<Expression>

Constructs a guard from the current state of the condition stack.

pub(crate) fn fold_guards(
&mut self,
prefix: &str,
guards: Vec<(Option<Expression>, Expression)>
) -> (Expression, Vec<Statement>)

Fold guards and expressions into a single expression. Note that this function assumes that at least one guard is present.

pub(crate) fn lookup_struct_symbol(
&self,
expression: &Expression
) -> Option<Symbol>

Looks up the name of the struct associated with an identifier or access expression, if it exists.

pub(crate) fn update_structs(&mut self, lhs: &Identifier, rhs: &Expression)

Updates self.structs for new assignment statements. Expects the left hand side of the assignment to be an identifier.

pub(crate) fn unique_simple_assign_statement(
&mut self,
expr: Expression
) -> (Identifier, Statement)

A wrapper around assigner.unique_simple_assign_statement that updates self.structs.

pub(crate) fn simple_assign_statement(
&mut self,
lhs: Identifier,
rhs: Expression
) -> Statement

A wrapper around assigner.simple_assign_statement that updates self.structs.

pub(crate) fn fold_returns(
&mut self,
block: &mut Block,
returns: Vec<(Option<Expression>, ReturnStatement)>
)

Folds a list of return statements into a single return statement and adds the produced statements to the block.