Struct leo_passes::loop_unrolling::unroller::Unroller

pub struct Unroller<'a> {
    pub(crate) symbol_table: RefCell<SymbolTable>,
    pub(crate) scope_index: usize,
    pub(crate) handler: &'a Handler,
    pub(crate) is_unrolling: bool,
}

Fields
symbol_table: RefCell<SymbolTable>

The symbol table for the function being processed.

scope_index: usize

The index of the current scope.

handler: &'a Handler

An error handler used for any errors found during unrolling.

is_unrolling: bool

Are we in the midst of unrolling a loop?

Implementations

impl<'a> Unroller<'a>

pub(crate) fn new(symbol_table: SymbolTable, handler: &'a Handler) -> Self

pub(crate) fn current_scope_index(&mut self) -> usize

Returns the index of the current scope. Note that if we are in the midst of unrolling an IterationStatement, a new scope is created.

pub(crate) fn enter_scope(&mut self, index: usize) -> usize

Enters a child scope.

pub(crate) fn exit_scope(&mut self, index: usize)

Exits the current block scope.

pub(crate) fn unroll_iteration_statement<I: LoopBound>(
&mut self,
input: IterationStatement,
start: Value,
stop: Value
) -> Statement

Unrolls an IterationStatement.

fn unroll_single_iteration<I: LoopBound>(
&mut self,
input: &IterationStatement,
iteration_count: I
) -> Statement

A helper function to unroll a single iteration an IterationStatement.