Struct leo_passes::loop_unrolling::unroller::Unroller

pub struct Unroller<'a> {
    pub(crate) constant_propagation_table: RefCell<ConstantPropagationTable>,
    pub(crate) symbol_table: RefCell<SymbolTable>,
    pub(crate) type_table: &'a TypeTable,
    pub(crate) scope_index: usize,
    pub(crate) handler: &'a Handler,
    pub(crate) node_builder: &'a NodeBuilder,
    pub(crate) is_unrolling: bool,
    pub(crate) current_program: Option<Symbol>,
}

Fields
constant_propagation_table: RefCell<ConstantPropagationTable>

A table of constant variables.

symbol_table: RefCell<SymbolTable>

The symbol table for the function being processed.

type_table: &'a TypeTable

A mapping from node IDs to their types.

scope_index: usize

The index of the current scope.

handler: &'a Handler

An error handler used for any errors found during unrolling.

node_builder: &'a NodeBuilder

A counter used to generate unique node IDs.

is_unrolling: bool

Are we in the midst of unrolling a loop?

current_program: Option<Symbol>

The current program name.

Implementations

impl<'a> Unroller<'a>

pub(crate) fn new( symbol_table: SymbolTable, type_table: &'a TypeTable, handler: &'a Handler, node_builder: &'a NodeBuilder ) -> Self

pub(crate) fn current_scope_index(&mut self) -> usize

Returns the index of the current scope. Note that if we are in the midst of unrolling an IterationStatement, a new scope is created.

pub(crate) fn enter_scope(&mut self, index: usize) -> usize

Enters a child scope.

pub(crate) fn exit_scope(&mut self, index: usize)

Exits the current block scope.

pub(crate) fn emit_err(&self, err: LoopUnrollerError)

Emits a Loop Unrolling Error

pub(crate) fn unroll_iteration_statement<I: LoopBound>( &mut self, input: IterationStatement ) -> Statement

Unrolls an IterationStatement.

fn unroll_single_iteration<I: LoopBound>( &mut self, input: &IterationStatement, iteration_count: I ) -> Statement

A helper function to unroll a single iteration an IterationStatement.