Struct leo_passes::static_single_assignment::static_single_assigner::StaticSingleAssigner

pub struct StaticSingleAssigner<'a> {
    pub(crate) rename_table: RenameTable,
    pub(crate) _handler: &'a Handler,
    pub(crate) counter: usize,
    pub(crate) is_lhs: bool,
    pub(crate) phi_functions: Vec<Statement>,
    pub(crate) condition_stack: Vec<Expression>,
    pub(crate) early_returns: Vec<(Option<Expression>, Expression)>,
}

Fields
rename_table: RenameTable

The RenameTable for the current basic block in the AST

_handler: &'a Handler

An error handler used for any errors found during unrolling.

counter: usize

A strictly increasing counter, used to ensure that new variable names are unique.

is_lhs: bool

A flag to determine whether or not the traversal is on the left-hand side of a definition or an assignment.

phi_functions: Vec<Statement>

Phi functions produced by static single assignment.

condition_stack: Vec<Expression>

A stack of condition Expressions visited up to the current point in the AST.

early_returns: Vec<(Option<Expression>, Expression)>

A list containing tuples of guards and expressions associated with early ReturnStatements.

Implementations

impl<'a> StaticSingleAssigner<'a>

pub(crate) fn new(handler: &'a Handler) -> Self

pub(crate) fn unique_symbol(&mut self, arg: impl Display) -> Symbol

Return a new unique Symbol from a &str.

pub(crate) fn simple_assign_statement(
place: Expression,
value: Expression
) -> Statement

Constructs the assignment statement place = expr;.

pub(crate) fn clear_phi_functions(&mut self) -> Vec<Statement>

Clears the self.phi_functions, returning the ones that were previously produced.

pub(crate) fn clear_early_returns(
&mut self
) -> Vec<(Option<Expression>, Expression)>

Clears the state associated with ReturnStatements, returning the ones that were previously produced.

pub(crate) fn push(&mut self)

Pushes a new scope, setting the current scope as the new scope’s parent.

pub(crate) fn pop(&mut self) -> RenameTable

If the RenameTable has a parent, then self.rename_table is set to the parent, otherwise it is set to a default RenameTable.

pub(crate) fn flatten_conditional_statement(
&mut self,
conditional_statement: ConditionalStatement
) -> Vec<Statement>

Introduces a new AssignStatement for non-trivial expressions in the condition of ConditionalStatements. For example,

if x > 0 { x = x + 1 } becomes let $cond$0 = x > 0; if $cond$0 { x = x + 1; }

if true { x = x + 1 } remains the same.

if b { x = x + 1 } remains the same. And then reconstructs and flattens ConditionalStatement.