Struct leo_passes::static_single_assignment::static_single_assigner::StaticSingleAssigner

pub struct StaticSingleAssigner<'a> {
    pub(crate) node_builder: &'a NodeBuilder,
    pub(crate) symbol_table: &'a SymbolTable,
    pub(crate) type_table: &'a TypeTable,
    pub(crate) rename_table: RenameTable,
    pub(crate) is_lhs: bool,
    pub(crate) assigner: &'a Assigner,
    pub(crate) program: Option<Symbol>,
}

Fields
node_builder: &'a NodeBuilder

A counter used to generate unique node IDs.

symbol_table: &'a SymbolTable

The SymbolTable of the program.

type_table: &'a TypeTable

A mapping from node IDs to their types.

rename_table: RenameTable

The RenameTable for the current basic block in the AST

is_lhs: bool

A flag to determine whether or not the traversal is on the left-hand side of a definition or an assignment.

assigner: &'a Assigner

A struct used to construct (unique) assignment statements.

program: Option<Symbol>

The main program name.

Implementations

impl<'a> StaticSingleAssigner<'a>

pub(crate) fn new( node_builder: &'a NodeBuilder, symbol_table: &'a SymbolTable, type_table: &'a TypeTable, assigner: &'a Assigner, ) -> Self

Initializes a new StaticSingleAssigner with an empty RenameTable.

pub(crate) fn push(&mut self)

Pushes a new scope, setting the current scope as the new scope’s parent.

pub(crate) fn pop(&mut self) -> RenameTable

If the RenameTable has a parent, then self.rename_table is set to the parent, otherwise it is set to a default RenameTable.

pub(crate) fn simple_assign_statement( &mut self, identifier: Identifier, rhs: Expression, ) -> Statement

pub(crate) fn unique_simple_assign_statement( &mut self, expr: Expression, ) -> (Identifier, Statement)

Constructs a simple assign statement for expr with a unique name. For example, expr is transformed into $var$0 = expr;. The lhs is guaranteed to be unique with respect to the Assigner.