Struct leo_passes::type_checking::checker::TypeChecker

pub struct TypeChecker<'a> {
    pub(crate) symbol_table: RefCell<SymbolTable>,
    pub(crate) type_table: &'a TypeTable,
    pub(crate) struct_graph: StructGraph,
    pub(crate) call_graph: CallGraph,
    pub(crate) handler: &'a Handler,
    pub(crate) function: Option<Symbol>,
    pub(crate) variant: Option<Variant>,
    pub(crate) has_return: bool,
    pub(crate) has_finalize: bool,
    pub(crate) is_finalize: bool,
    pub(crate) is_imported: bool,
    pub(crate) is_return: bool,
}

Fields
symbol_table: RefCell<SymbolTable>

The symbol table for the program.

type_table: &'a TypeTable

A mapping from node IDs to their types.

struct_graph: StructGraph

A dependency graph of the structs in program.

call_graph: CallGraph

The call graph for the program.

handler: &'a Handler

The error handler.

function: Option<Symbol>

The name of the function that we are currently traversing.

variant: Option<Variant>

The variant of the function that we are currently traversing.

has_return: bool

Whether or not the function that we are currently traversing has a return statement.

has_finalize: bool

Whether or not the function that we are currently traversing invokes the finalize block.

is_finalize: bool

Whether or not we are currently traversing a finalize block.

is_imported: bool

Whether or not we are currently traversing an imported program.

is_return: bool

Whether or not we are currently traversing a return statement.

Implementations

impl<'a> TypeChecker<'a>

pub fn new( symbol_table: SymbolTable, type_table: &'a TypeTable, handler: &'a Handler ) -> Self

Returns a new type checker given a symbol table and error handler.

pub(crate) fn enter_scope(&mut self, index: usize)

Enters a child scope.

pub(crate) fn create_child_scope(&mut self) -> usize

Creates a new child scope.

pub(crate) fn exit_scope(&mut self, index: usize)

Exits the current scope.

pub(crate) fn emit_err(&self, err: TypeCheckerError)

Emits a type checker error.

fn check_type( &self, is_valid: impl Fn(&Type) -> bool, error_string: String, type_: &Option<Type>, span: Span )

Emits an error to the handler if the given type is invalid.

pub(crate) fn check_eq_types( &self, t1: &Option<Type>, t2: &Option<Type>, span: Span )

Emits an error if the two given types are not equal.

pub(crate) fn assert_and_return_type( &self, actual: Type, expected: &Option<Type>, span: Span ) -> Type

Use this method when you know the actual type. Emits an error to the handler if the actual type is not equal to the expected type.

pub(crate) fn assert_type( &self, actual: &Option<Type>, expected: &Type, span: Span )

Emits an error to the error handler if the actual type is not equal to the expected type.

pub(crate) fn assert_address_type(&self, type_: &Option<Type>, span: Span)

Emits an error to the error handler if the given type is not an address.

pub(crate) fn assert_bool_type(&self, type_: &Option<Type>, span: Span)

Emits an error to the handler if the given type is not a boolean.

pub(crate) fn assert_field_type(&self, type_: &Option<Type>, span: Span)

Emits an error to the handler if the given type is not a field.

pub(crate) fn assert_group_type(&self, type_: &Option<Type>, span: Span)

Emits an error to the handler if the given type is not a group.

pub(crate) fn assert_scalar_type(&self, type_: &Option<Type>, span: Span)

Emits an error to the handler if the given type is not a scalar.

pub(crate) fn assert_signature_type(&self, type_: &Option<Type>, span: Span)

Emits an error to the handler if the given type is not a signature.

pub(crate) fn assert_int_type(&self, type_: &Option<Type>, span: Span)

Emits an error to the handler if the given type is not an integer.

pub(crate) fn assert_signed_int_type(&self, type_: &Option<Type>, span: Span)

Emits an error to the handler if the given type is not a signed integer.

pub(crate) fn assert_unsigned_int_type(&self, type_: &Option<Type>, span: Span)

Emits an error to the handler if the given type is not an unsigned integer.

pub(crate) fn assert_magnitude_type(&self, type_: &Option<Type>, span: Span)

Emits an error to the handler if the given type is not a magnitude (u8, u16, u32).

pub(crate) fn assert_bool_int_type(&self, type_: &Option<Type>, span: Span)

Emits an error to the handler if the given type is not a boolean or an integer.

pub(crate) fn assert_field_int_type(&self, type_: &Option<Type>, span: Span)

Emits an error to the handler if the given type is not a field or integer.

pub(crate) fn assert_field_group_type(&self, type_: &Option<Type>, span: Span)

Emits an error to the handler if the given type is not a field or group.

pub(crate) fn assert_field_group_int_type( &self, type_: &Option<Type>, span: Span )

Emits an error to the handler if the given type is not a field, group, or integer.

pub(crate) fn assert_field_group_signed_int_type( &self, type_: &Option<Type>, span: Span )

Emits an error to the handler if the given type is not a field, group, or signed integer.

pub(crate) fn assert_field_scalar_int_type( &self, type_: &Option<Type>, span: Span )

Emits an error to the handler if the given type is not a field, scalar, or integer.

pub(crate) fn assert_field_group_scalar_int_type( &self, type_: &Option<Type>, span: Span )

Emits an error to the handler if the given type is not a field, group, scalar, integer, or boolean.

pub(crate) fn assert_castable_type(&self, type_: &Option<Type>, span: Span)

Emits an error to the handler if the given type is not a field, group, scalar, integer, boolean, or address.

pub(crate) fn get_core_constant( &self, type_: &Type, constant: &Identifier ) -> Option<CoreConstant>

Type checks the inputs to an associated constant and returns the expected output type.

pub(crate) fn get_core_function_call( &self, struct_: &Type, function: &Identifier ) -> Option<CoreFunction>

Emits an error if the struct is not a core library struct. Emits an error if the function is not supported by the struct.

pub(crate) fn check_core_function_call( &self, core_function: CoreFunction, arguments: &[(Option<Type>, Span)], function_span: Span ) -> Option<Type>

Type checks the inputs to a core function call and returns the expected output type. Emits an error if the correct number of arguments are not provided. Emits an error if the arguments are not of the correct type.

pub(crate) fn check_expected_struct( &mut self, struct_: Identifier, expected: &Option<Type>, span: Span ) -> Type

Returns the struct type and emits an error if the expected type does not match.

pub(crate) fn assert_member_is_not_record( &self, span: Span, parent: Symbol, type_: &Type )

Emits an error if the struct member is a record type.

pub(crate) fn assert_type_is_valid(&self, type_: &Type, span: Span) -> bool

Emits an error if the type or its constituent types is not valid.

pub(crate) fn assert_mapping_type( &self, type_: &Option<Type>, span: Span ) -> Option<MappingType>

Emits an error if the type is not a mapping.

pub(crate) fn assert_array_type(&self, type_: &Option<Type>, span: Span)

Emits an error if the type is not an array.